Fix issue with route still returning none
