+++ blockchain/my_id.txt
@@ -1,1 +1,1 @@
-NONAME
+VERSILISTYSON
