+++ hashtables_sprint/ex1/README.md
