+++ hashtables_sprint/ex1/hashtables_impl.py
