+++ hashtables_sprint/ex2/hashtables.py
